rename tests to unit tests
